changed variable name

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t s = 0;
+	int s = 0;
 
 	while (h != NULL)
 	{
